+++ pgqueue/postgresql_queue.go
@@ -22,8 +22,8 @@ func (queue *PostgreSQLQueue) Push(ctx context.Context, querier Querier, newJob 
 		(id, created_at, updated_at, scheduled_for, failed_attempts, status, type, data, retry_max, retry_delay, retry_strategy, timeout)
 		VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12)`, tableName),
 		job.ID, job.CreatedAt, job.UpdatedAt, job.ScheduledFor, job.FailedAttempts,
-		int32(job.Status), job.Type, job.DataJson, job.RetryMax, job.RetryDelay,
-		int32(job.RetryStrategy), job.Timeout)
+		job.Status, job.Type, job.DataJson, job.RetryMax, job.RetryDelay,
+		job.RetryStrategy, job.Timeout)
 	if err != nil {
 		return fmt.Errorf("pgqueue: inserting job: %w", err)
 	}
@@ -45,8 +45,8 @@ func (queue *PostgreSQLQueue) PushMany(ctx context.Context, querier Querier, new
 			return fmt.Errorf("pgqueue: validating job %d: %w", i, err)
 		}
 		rows[i] = []any{
-			job.ID, job.CreatedAt, job.UpdatedAt, job.ScheduledFor, job.FailedAttempts, int32(job.Status),
-			job.Type, job.DataJson, job.RetryMax, job.RetryDelay, int32(job.RetryStrategy), job.Timeout,
+			job.ID, job.CreatedAt, job.UpdatedAt, job.ScheduledFor, job.FailedAttempts, job.Status,
+			job.Type, job.DataJson, job.RetryMax, job.RetryDelay, job.RetryStrategy, job.Timeout,
 		}
 	}
 
@@ -92,7 +92,7 @@ func (q *PostgreSQLQueue) Pull(ctx context.Context, numberOfJobs uint64) ([]Job,
 		}
 
 		rows, err := q.pool.Query(ctx, query,
-			int32(JobStatusRunning), now, int32(JobStatusQueued), now, int64(numberOfJobs))
+			JobStatusRunning, now, JobStatusQueued, now, int64(numberOfJobs))
 		if err != nil {
 			return nil, fmt.Errorf("pgqueue: pulling jobs: %w", err)
 		}
@@ -137,7 +137,7 @@ func (q *PostgreSQLQueue) FailJob(ctx context.Context, jobID uuid.UUID) error {
 			ELSE $3 + (retry_delay * CASE WHEN retry_strategy = $4 THEN failed_attempts + 1 ELSE 1 END) * INTERVAL '1 second'
 		END
 	WHERE id = $5`, tableName),
-		int32(JobStatusFailed), int32(JobStatusQueued), now, int32(RetryStrategyExponential), jobID)
+		JobStatusFailed, JobStatusQueued, now, RetryStrategyExponential, jobID)
 	if err != nil {
 		return fmt.Errorf("pgqueue: failing job: %w", err)
 	}
@@ -166,7 +166,7 @@ func (q *PostgreSQLQueue) GetJob(ctx context.Context, jobID uuid.UUID) (Job, err
 func (q *PostgreSQLQueue) GetFailedJobs(ctx context.Context, limit int64) ([]Job, error) {
 	rows, err := q.pool.Query(ctx, fmt.Sprintf(`SELECT * FROM %s WHERE status = $1 ORDER BY created_at DESC LIMIT $2`,
 		pgx.Identifier{q.table}.Sanitize()),
-		int32(JobStatusFailed), limit)
+		JobStatusFailed, limit)
 	if err != nil {
 		return nil, fmt.Errorf("pgqueue: querying failed jobs: %w", err)
 	}
